requirement.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dlib==19.18.0
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.0.4
gTTS-token==1.1.3
idna==2.8
imutils==0.5.3
joblib==0.14.0
numpy==1.17.4
opencv-contrib-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
pydub==0.23.1
pygame==1.9.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
soupsieve==1.9.5
tkMagicGrid==1.0.3
urllib3==1.25.7
